requirements.txt
asgiref==3.7.2
Django==4.2.6
django-ninja==0.22.2
gunicorn==21.2.0
packaging==23.2
pycryptodome==3.19.0
pydantic==1.10.13
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.8.0
